pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image_processing"
version = "0.1.0"
description = "Image processing application"
requires-python = ">=3.10"
dependencies = [
    "mypy>=1.19.1",
    "pika>=1.3.2",
    "poethepoet>=0.41.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "poethepoet>=0.24",
    "ruff>=0.4",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.poe.tasks]
test = "pytest"
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy src main.py"
